fix: check user rights to update or delete mitt in service

+++ internal/service/mitt/mitt.go
@@ -66,7 +66,20 @@ func (s *Service) GetAllUserMitts(ctx context.Context, userID uuid.UUID) ([]*mod
 	return mitts, nil
 }
 
-func (s *Service) UpdateMitt(ctx context.Context, mittID uuid.UUID, mitt *models.MittUpdate) (*models.Mitt, *models.HTTPError) {
+func (s *Service) UpdateMitt(ctx context.Context, userID uuid.UUID, mittID uuid.UUID, mitt *models.MittUpdate) (*models.Mitt, *models.HTTPError) {
+	existingMitt, httpErr := s.GetMitt(ctx, mittID)
+	if httpErr != nil {
+		return nil, httpErr
+	}
+
+	// Check if user is author of mitt
+	if existingMitt.Author != userID {
+		return nil, &models.HTTPError{
+			Code:    http.StatusForbidden,
+			Message: "You are not allowed to do this",
+		}
+	}
+
 	newMitt, err := s.mr.UpdateMitt(ctx, mittID, mitt)
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
@@ -84,7 +97,20 @@ func (s *Service) UpdateMitt(ctx context.Context, mittID uuid.UUID, mitt *models
 	return newMitt, nil
 }
 
-func (s *Service) DeleteMitt(ctx context.Context, mittID uuid.UUID) *models.HTTPError {
+func (s *Service) DeleteMitt(ctx context.Context, userID uuid.UUID, mittID uuid.UUID) *models.HTTPError {
+	existingMitt, httpErr := s.GetMitt(ctx, mittID)
+	if httpErr != nil {
+		return httpErr
+	}
+
+	// Check if user is author of mitt
+	if existingMitt.Author != userID {
+		return &models.HTTPError{
+			Code:    http.StatusForbidden,
+			Message: "You are not allowed to do this",
+		}
+	}
+
 	err := s.mr.DeleteMitt(ctx, mittID)
 	if err != nil {
 		slog.Error("error deleting mitt", slog.Any("err", err))
